Move csv_utils.go to the e2etests package
